fix: refactor gradle resolution into versionsresolver
